+++ Backend/EvitarBackEnd/Models/MovEPIModelView.cs
@@ -5,9 +5,9 @@ namespace EvitarBackEnd.Models
 {
     public class MovEPIModelView{
         [Key]
-        public int IdMovimento {get;set;}
+        public long IdMovimento {get;set;}
         [Key]
-         public int IdEPI {get;set;}
+         public long IdEPI {get;set;}
          public string TypeMov {get;set;}
         public int IdColaborador {get;set;}
         public int Check {get;set;}
