+++ Backend/EvitarBackEnd/Models/RegisterModel.cs
@@ -10,7 +10,7 @@ namespace EvitarBackEnd.Models.Users
         [Required]
         public string Password { get; set; }
 
-        public int IdColaborador{get;set;}
+        public long IdColaborador{get;set;}
         [System.ComponentModel.DataAnnotations.Schema.ForeignKey("IdColaborador")]
         public ColaboradorModel IdColaboradorForeignKey {get;set;}
     }
