+++ Backend/EvitarBackEnd/Models/UpdateModel.cs
@@ -7,7 +7,7 @@ namespace EvitarBackEnd.Models.Users
         public string Username { get; set; }
         public string Password { get; set; }
 
-        public int IdColaborador{get;set;}
+        public long IdColaborador{get;set;}
         [System.ComponentModel.DataAnnotations.Schema.ForeignKey("IdColaborador")]
         public ColaboradorModel IdColaboradorForeignKey {get;set;}
     }
